Add top10 duck
