+++ src/ducks/index.js
@@ -1,18 +1,20 @@
 import { combineReducers } from 'redux-immutable'
-import recent from './recent'
+import analyzer from './routes'
+import flight from './flight'
 import general from './general'
 import liveflights from './liveflights'
+import recent from './recent'
 import search from './search'
-import flight from './flight'
-import analyzer from './routes'
+import top10 from './top10'
 
 const app = combineReducers({
+  analyzer,
   flight,
   general,
   liveflights,
   recent,
-  analyzer,
-  search
+  search,
+  top10
 })
 
 export default app
